pyproject.toml
[project]
name = "autogen-fastapi-postgres"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "autogen_agentchat",
    "autogen-ext[openai]",
    "fastapi",
    "uvicorn",
    "PyYAML",
    "streamlit",
    "sqlalchemy>=2.0.0",
    "psycopg2-binary",
    "alembic",
    "asyncpg",
    "greenlet",
    "python-dotenv",
    "chromadb",
    "sentence-transformers"
]

[tool.alembic]
script_location = "alembic"

[tool.setuptools.packages.find]
include = ["agentchat_fastapi*", "agentchat_chainlit*", "agentchat_streamlit*", "core_semantic_router*", "autogenchat_bi*"]